+++ CHANGELOG.md
@@ -3,12 +3,15 @@
 ### HEAD
 
 ```
+```
+
+### 0.3.0 (2014-03-19)
+
 * Update for flags, options, examples
 * Fix for duplicated client names and groups (issue #8)
 * New option for client command execution method (-ccem)
 * Worker pkg - worker.go, cce.go
 * Parallel command execution (issue #5)
-```
 
 ### 0.2.5 (2014-03-16)